Style css (#1)
* Meteor built

* Dynamic Templates

This has only been partially implemented, it does not contain full list of options.  More templates are needed, styling is incomplete.

* add style to drill disaster alerts

* Adding alert and summary

This is a conversion of Kalen's work being merged with Yizan's.  It is not yet complete.

* Cleanup

A few inconsequential changes to improve readability.  Separated to avoid confusion in later commits.

* Filling in some blanks

Added in a confirmation after the summary that I had in my emailed file. Filled out the buttons that all said "missile." Still working on it right now.

* Interface v1

Initial, hard-coded version of the interface.  Will change to be dynamic later.

* User sees if they canceled

When they go back to first screen they'll see whether or not they canceled.

* README