Fix device
The problem was that Session did not import the device files, so when it was retrieved, it no longer had associated functions.  To solve this, the device has been removed from Session.  This means reloading the page will generate a new device.  However, since  the message is set immediately before being sent, this should not be a problem.