Finally found some way around the Session variable
So I made changes to the TV_driver.js and TV.js file to allow it to run the one test I made. Basically the import session and the use of Session.get makes it so that they file can't be used outside of Meteor. To get around this, I modified two of the functions to not use Session.get() (it appears to still work fine, I think using it more than once might be duplicate work), and when testing I comment out the import session line (as we're just testing the output of the function I feel this should be alright as the Session object has no bearing on the input and output of the function if the Session object isn't used).

The test is run on the from the node REPL. All the files have comments explaining what it is. Note that Meteor and the node.js repl uses different syntax and a couple lines needs to be changed when moving back and forth between the application and test.